drivers: debug option to disable --gpgpu
For server installs, we specifically pass "--gpgpu" to ubuntu-drivers
calls, but this stops broadcom-sta-dkms drivers from being
a candidate to install. It's faster to test on server, so we can
enable this flag to test installs of braodcom-sta-dkms on server
(which normally shouldn't happen).

Renames the `is_server` variable to `use_gpgpu` to be more clear
about the intent.