remove reference to cloudinit.safeyaml.dumps
With the goal of removing cloud-init from the snap, we can reimplement
cloudinit.safeyaml.dumps in subiquity. safeyaml.dumps is just a wrapper
script around yaml.dump [1] and we only use reference it once using its
defaults. Replace the existing call with the equivalent yaml.dump
invocation and remove the cloudinit import.

[1] https://github.com/canonical/cloud-init/blob/6e4153b346bc0d3f3422c01a3f93ecfb28269da2/cloudinit/safeyaml.py#L261