controller: fix _active
If interactively installing on a variant that doesn't require identity,
there was a nasty problem.  The previous logic reevaluated part of the
interactive state based on the variant, but only after confirmation was
set.

The middleware check looking to do confirmation didn't quite work,
because the "identity" model wasn't a postinstall_only model (or at all)
but didn't get caught by the earlier skip code because the full answer
for interactive was calculated after confirmation.

To break that dependency loop, make it so we can do the variant check
sooner.