storage: set _netdev option more consistently on mounts
Instead of expecting the caller of add_mount() to consistently pass a
"on_remote_storage" argument, we now let the function inspect the
attributes of the associated file-system. If the file-system appears to
be backed on remote storage, then the _netdev option is added ;
otherwise it is not.

This makes the code simpler and fixes the absence of _netdev when using
a NVMe/TCP drive in guided.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>