tui: avoid redrawing the screen if there is no screen
When a debug shell is running (or if Subiquity is shutting down),
calling urwid_loop.draw_screen() will cause urwid to raise an
AssertionError(self._started). Since we are scheduling calls to
redraw_screen using asyncio, we need to ensure we skip the call to
urwid_loop.draw_screen() if now is not the time.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>