storage: add a refresh and back buttons when no disk are discovered
When no disk are discovered by the storage probing operation, Subiquity
used to show a screen that would not allow the user to go back or do
anything useful. The screen states something along the lines of "no
disks were detected, the installation will not be impossible".

Unfortunately, this means that devices that have been discovered lately
cannot be used without restarting the installer. Such devices would
include devices that have been hot-plugged (which is rather uncommon on
Ubuntu Server) and network storage devices such as NVMe/TCP drives.

Instead of preventing the user to go back, we now show a "Back" button,
allowing them to re-enter the storage screen with newly probed disks. We
also add a "Refresh" button, which essentially does the does the same as
re-entering the screen without going back.

This way, in the NVMe/TCP use-case, if we get to the storage screen
before the NVMe drives are connected, we do not need to restart the
installer. We can open a shell, run the nvme connect-all commands ; and
then hit Refresh.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>