error: fix crash when showing crash report early
When showing a crash report very early, Subiquity will often fail with
the following exception:

  File "subiquity/ui/views/error.py", line 221, in __init__
    (w, self.pile.options("pack")) for w in self._pile_elements()
                                            ^^^^^^^^^^^^^^^^^^^^^
  File "subiquity/ui/views/error.py", line 260, in _pile_elements
    assert self.report
           ^^^^^^^^^^^
AssertionError
make: *** [Makefile:41: dryrun] Error 1

This happens when the report has finished generating on the server-side
but has not been loaded yet on the client-side. In that scenario, the
meta/status endpoint returns ErrorReportState.DONE ; but we need to wait
until the report is loaded on the client-side before using it.

Fixed by showning a "Report is loading" dialog until the report has
finished loading client-side.

LP: #2076914

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>