storage: lean on systemd-shutdown to export zpools on shutdown
People have reported issues involving non exported zpools at the end of
Ubuntu installations. Diagnosing such issues is tricky because the
`zpool export` commands have to run after the target is unmounted. And
once the target is unmounted, logs can no longer be copied over to the
target so bug reports don't include traces of potential `zpool export`
failures (or failures to unmount target).

Instead of relying on Subiquity's _pre_shutdown hook, we now make use of
systemd-shutdown to export the zpools. The chance of successfully
exporting the zpools is expected to be higher since it will happen later
in the shutdown sequence ; after more mounts have been detached.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>