ui: replace use of deprecated user_arg by kw-only arg user_args
urwid deprecated the user of the user_arg parameter of connect_signal
(and might get dropped sometime).  Instead, the suggested approach is to
use user_args (plural!) as a kw-only argument.

We have to change the signature of callbacks functions because
`user_args` is passed as an argument at the beginning ; whereas
`user_arg` is passed as an argument at the end.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>