Don't attempt to enumerate over an empty command list
This avoids a crash in both the installer and the validation script when
a command list such as early-commands is provided in the Autoinstall
YAML, but no entries are included. This causes the value of that key to
be None instead of an empty list and triggers an exception in
load_autoinstall_data(). Ignoring that value causes it to behave as if
that command list was not provided at all.