ci: allow null compare in check-yaml-fields script
Allow check for `None` type in check-yaml-fields.py script. It wasn't
clear from the original commit - 2ddde49a4 - if we purposefully didn't
want to compare `null` in the script, or it was just a typo. This will
now let us assert a field should be `null` instead of some other value,
and fail properly when it is not.