apt: fix invocation of setup_overlay in dry-run mode
setup_overlay() expects a list of lower directories. That said, we have
one invocation in dry-run mode where we pass it a path (which is a
string) directly instead of wrapping the path in a list with one
element.

The dry-run implementation of setup_overlay() accesses the first element
of the list using lowers[0], and expects a path out of it. But when
lowers is a string, lowers[0] returns the first character of string,
which usually is "/" and therefore corresponds to a valid path.

It makes the code accidentally work, but use a path that is not the one
we expected.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>