Fix mount fstype parameter
Currently the fstype parameter is not passed when its not associated
with a device.

This fix aims to pass the fstype mount parameter when there is no device
associated with the mount, this is needed when defining bind-mounts for
example. We also add an exception for ZFS, since there is no device
associated with a ZFS mount.

Signed-off-by: David Negreira <david.negreira@canonical.com>