storage: in the presence of a NBFT, assume the firmware supports NVMe/TCP
Checking for known-good firmware models/versions is one way to determine
if the system supports NVMe/TCP booting. That said, if a NBFT is present
on the system, it is a much stronger hint that NVMe/TCP booting will be
possible.

We can't completely replace the mechanism because sometimes, the NBFT
will not be there during installation ; even though the system can boot
using NVMe/TCP.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>