cloud: refactor cloud-init validation
Extract cloud-init schema validation to a new system_script
`subiquity-legacy-cloud-init-validate`. This is called "legacy" because,
in the future, we likely want to rely on the cloud-init CLI directly to
do schema validation, but there isn't a pressing need to rely on it now.
For now, all releases will rely on the "legacy" script.

Also let `get_unknown_keys` use `system_scripts_env` for the command
environment to rely on the system cloud-init version.