cloud: top-level key validation skip check
The "schema" subcommand was introduced in cloud-init 22.2 (see
cloud-init commit [0] ), which was released April 27, 2022. This means
older releases of Focal and Jammy, which are currently supported LTSes,
will fail this code path after we unstage cloud-init from the Subiquity
snap. Add a check to skip this validation if a newer version of
cloud-init is detected.

[0] 3bcffacb216d683241cf955e4f7f3e89431c1491