filesystem: only execute zpool export -a if zpool is installed
We recently made the call to `zpool export -a` unconditional to account
for "advanced" autoinstall scenarios. However, the zpool command is not
always present in the live installer environment ; and it results in a
Subiquity crash if the command fails to execute. Today, zpool (provided
by zfsutils-linux) is dynamically installed in the live environment by
curtin when it handles a ZFS installation.

Let's make sure we don't crash if zpool is not available when shutting
down. It just means we're not using ZFS.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>