kvm-test: set cloud-config disk after real disks
Ensure that the disks used for installation are passed to kvm first
and not the disk with cloud-config/autoinstall data. Sometimes
curtin will put paths to disk in the fstab like:

    /dev/disk/by-path/pci-000n-xxxx/by-uuid/UUID

and when we reboot without the cloud-config disk the path to the
install disk will change since the number of pci devices connected
to the machine has technically changed. We can make this a little more
stable by letting the cloud-config drive be the last drive attached.