ui: fix journal parsing in ProgressView
The ProgressView.other_event function splits messages on the ":"
character to get the context type and the message from journal messages.
This code will crash if the message has more than one ":" character.
When passing cloud-config with invalid top level keys, subiquity will
check these keys and report on the journal with a message:
"cloud-init schema validation failure for: <key1> [, <key2>, ...]",
which was causing the UI to crash repeatedly.

LP: #2080918