storage: don't produce target-resize if we can't fit more partitions
Just like we did for use-gap scenarios, we now make sure we have enough
room for primary partitions when suggesting a target-resize scenario.

We need 1 primary partition to store the new Ubuntu system (unless we're
making room inside the extended) and potentially 1 more to satisfy the
boot plan (e.g., creation of an ESP).

Recovery partitions and capabilities having different requirements than
DIRECT are not accounted for.

LP: #2081724

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>