utils: introduce passlib
The `crypt` module was deprecated in python3.11 and officially removed
in python3.13. This commit introduces equivalent functionality using
the `passlib` module instead. The `passlib` module is listed directly in
the python documentation as a replacement for the `crypt` module[0].
Additionally, this module is already provided by the `python3-passlib`
package in main.

The main functionality of `crypt_password` is not replaced in this commit.
Instead, this commit acts more as a transient commit to demonstrate
that, given the right parameters, we can use `passlib` to provide
equivalent functionality compared with what we expected from the
`crypt`-based implementation.

[0] https://docs.python.org/3.12/library/crypt.html