utils: replace crypt with passlib
Replace usage of the `crypt` module from the stdlib with `passlib`,
which is provided by the python3-passlib (main) package. Follow-up
to the previous commit which introduced the `passlib`-based
functionality, and now switches `crypt_password` over to using it.

The equivalence between the two implementations can be further
demonstrated by the canary test passing with just the previous commit
and also passing with this one.