Fix switching back from console-conf to getty
When a password for a user is set, then we should use getty rather
than console-conf. Unit conditions did not work along conflicts,
because conflicts are resolved during loading of systemd, and
conditions are resolved right before executing the unit.

Instead we always use console-conf. If we find that we should have
switched to getty, then we return a special value that is marked as
not restarting (we always restart otherwise). An ExecStopPost also
matches that exit status to start getty (the default job mode is
"replace", which means the conflict will also kick console-conf out).