storage: look harder for the gap after removing a partition
When removing a partition, we expect the resulting gap to be either
before (if there was free space) or exactly where the partition used to
be.

However, in practice, Subiquity sometimes places the gap slightly after
the partition. This is probably a bug but for now let's look harder for
the resulting gap.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>