storage: do not renumber logical parts after removing a primary
The renumber_logical_partition function does not do the right thing if
called after removing a primary partition. As it is today, the
implementation does not ensure that the new partition numbers are in the
range of logical partition numbers.

As a consequence, if we remove partition 1 from a MSDOS partition
table, the logical partitions will be renumbered 2, 3, 4, ... which are
not in the 5+ range. The new numbers might also conflict with primary
partition 2, 3, and 4 if they exist.

Fixed by invoking the function only after removing a logical partition.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>
(cherry picked from commit ad92d7aa54640bb03eb3680db53cfb148f655fc4)