controller(server): update interactive function doc, signature, tests
Make the return value of `SubiquityController.interactive` an explicit
False return if the controller is not interactive.

Previously, if none of the `if` blocks passed then `interactive` would
implicility return `None`, which would effectively return `False` when
used in `if controller.interactive()` checks. Let's make it explicit.

Additionally, add a docstring that describes the ways in which the _active
value can be changed and generally when controllers get configured.

Lastly, add some tests to the controller class to assert behavior about
interactivity and the interactive_for_variants functionality.

Co-authored-by: Michael Hudson-Doyle <michael.hudson@canonical.com>