server: unify setting the variant
Fix a mismatch between the variant of the server and the variant of the
SubiquityModel. POST requests to /meta/client_variant would modify the
variant for both the server and the SubiquityModel, while POST requests to
the source controller would update only the SubiquityModel.

Fix this by making all requests to change the variant go through a
single function in the server and let that be responsible for updating
itself and the SubiquityModel.

(Conceptually the Source controller should not be responsible for
updating the SubiquityModel.)