test: update userdata test to use a real base_model instead of mock
By letting the UserdataController use a mocked base_app, the tests
load_none, load_empty, and load_some don't actually verify the
inputs and only test that the data is loaded into the model. In
particular, the load_some test was loading bad cloud-config but still
had a passing test. Change the base_app attribute to actually be
a SubiquityModel by default.

Additionally, update the load_some test to not use bad cloud-config and
add a test to make sure validation errors are thrown on bad config.