install: remove previous uuid= and rp-partuuid= in the factory reset grub config
We received a report that, after creating a USB reset media from the
system and reinstalled, the system cannot boot into factory reset
partition again.  We found out that there are two uuid= entries in
`etc/grub/grub.cfg` in the reinstalled factory reset partition.

This is to remove the additional uuid= and rp-partuuid= entries if
they are found in the kernel cmdline we are going to manipulate.

(cherry picked from commit 938c0edfd70982817f9c1e0f8d959aa8b6197c69)