storage: exclude erase-install scenarios that renumber partitions
The /storage/v2 API currently does not deal well with partitions
renumbering. For logical partitions, renumbering automatically occurs
when a partition is removed or added. This leads to inconsistencies and
mismatches in the info returned by the /storage/v2 API ; as reported in
https://launchpad.net/bugs/2091172

The result of the these inconsistencies gets reflected in the summary of
storage changes shown to the user in the UI. Having wrong information
there will scare people away.

For now, we will avoid producing guided erase and install scenarios if
it means renumbering partitions.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>