storage: ensure make_partition honors flag=logical when finding gap
In the past, make_partition would always determine the offset of the
new partition by looking for the largest gap on disk. However, for
logical partitions, this would return the wrong gap. This was breaking
some logic in tests since removing a logical partition created with
make_partition would produce a gap outside the extended partition -
where we would expect the opposite.

Fixed by finding the largest gap in the extended partition when
flag="logical" is passed.

As a bonus, we are not anymore forced to create the logical partitions
immediately after creating the extended partition.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>