storage: disable logical parts renumbering for erase-install
When doing erase-and-install, we should be able to skip logical
partitions renumbering altogether. Logical partitions renumbering is
important so that there is no "missing" partition number in an extended
partition.

However, for erase-and-install, the partition we remove creates space
that we will use to create one more partition. So the number of logical
partitions should not change.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>