storage: set backlinks only if other post-init methods succeed
When creating a fsobj, e.g., a partition, we automatically modify
associated fsobj-s (e.g., the underlying disk) to make them aware of the
newly created object.

We do this through the use of backlinks and more specifically the
function _set_backlinks. Currently, this function is automatically
called first when entering the fsobj's __post_init__ method.

If a fsobj defines its own __post_init__ method (let's called it
user-defined), it gets called after _set_backlinks. This is usually
fine.

However if the user-defined __post_init__ raises an exception, we end up
with backlinks referencing an object that was not successfully
initialized.

This is what happens with the Partition fsobj. In its user-defined
__post_init__, we determine the partition number by looking at the
underlying disk. If the number could not be determined (because there is
no more number available), we raise an exception "Exceeded number of
available partitions". However, at this point, the underlying disk is
already aware of the new partition - through the use of backlinks.

Iterating over the list of partitions will then raise a TypeError when
trying to use partition.number - which has not been successfully
assigned.

Fixed by calling _set_backlinks after the user-defined __post_init__, so
that we only set the backlinks on success.

refs LP: #2095159

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>