storage: prevent SystemGetter from cancelling and use a lock
The SystemGetter.get() method essentially does 3 important things:
 * it mounts stuff in /var/lib/snapd/seed
 * it queries snapd (and we expect snapd to read contents from /var/lib/snapd/seed)
 * it then tries to cleanup by unmounting what it previously mounted

Unfortunately, if we cancel this task in the middle, especially while it is
querying snapd, the cleanup code will fail to unmount things because snapd is
still using /var/lib/snapd/seed (cancelling the query does not mean that snapd
will immediately release the resources - or even stop processing the query).
Typically this results in a failure to unmount pc-kernel.snap with EBUSY.

Effectively, if SystemGetter.get() gets cancelled, we would need to wait until
the query to snapd finishes before we run the cleanup code.

But the approach that this patch takes is to make sure SystemGetter.get() is
not cancelled in the first place.

SystemGetter.get() was subject to cancellation because it runs as part of
_examine_systems - which is cancelled and restarted when the source changes.

We now make sure that we shield SystemGetter.get() from cancellation when we
execute it from _examine_systems. If _examine_system gets cancelled,
SystemGetter.get() will now continue running in the background.

This leads to another problem. If we cancel _examine_system and rerun it, we
will end up with two concurrent calls to SystemGetter.get() (the new one +
the one we want to finish). Obviously, things go sideways if this happens.

To avoid the problem, we now use the async_helpers.exclusive decorator for
SystemGetter.get() - so that only one task runs the critical section at a given
time.

LP: #2084032

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>