storage: prevent _get_system from cancelling and use a lock
The _get_system() method essentially does 3 important things:
 * it mounts stuff in /var/lib/snapd/seed
 * it queries snapd (and we expect snapd to read contents from /var/lib/snapd/seed)
 * it then tries to cleanup by unmounting what it previously mounted

Unfortunately, if we cancel this task in the middle, especially while it is
querying snapd, the cleanup code will fail to unmount things because snapd is
still using /var/lib/snapd/seed (cancelling the query does not mean that snapd
will immediately release the resources - or even stop processing the query).
Typically this results in a failure to unmount pc-kernel.snap with EBUSY.

Effectively, if _get_system() gets cancelled, we would need to wait until
the query to snapd finishes before we run the cleanup code.

But the approach that this patch takes is to make sure _get_system() is
not cancelled in the first place.

_get_system() was subject to cancellation because it runs as part of
_examine_systems - which is cancelled and restarted when the source changes.

We now make sure that we shield _get_system() from cancellation when we
execute it from _examine_systems. If _examine_system gets cancelled,
_get_system() will now continue running in the background.

This leads to another problem. If we cancel _examine_system and rerun it, we
will end up with two concurrent calls to _get_system() (the new one +
the one we want to finish). Obviously, things go sideways if this happens.

To avoid the problem, we now use the async_helpers.exclusive decorator for
_get_system() - so that only one task runs the critical section at a given
time.

LP: #2084032

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>
(cherry picked from commit 1f94353223dfb0cc942d9b24c5c8ea3e37846c5f)

Backported to use _get_system() instead of SystemGetter.get() from the original
patch.