storage: accept to specify the source-id
Internally, _get_system(variation) expects that the variation exists for
the currently configured source.

However, because of the lock we added to _get_system(variation), it is now
more likely that the configured source has changed by the time we start
mounting stuff and querying snapd.

When that happens, _get_system(variation) fails with KeyError ; because
the variation is no longer present in the currently configured source.

Fixed by making sure that we use the source handler that was configured when
_get_system(variation) was first invoked.

LP: #2084032

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>
(cherry picked from commit 65983be4308e302a3dfa48513507e562dd002669)

Backported to use _get_system() instead of SystemGetter.get() from the original
patch.