storage: only if it makes sense, cancel SystemGetter.get
Now that SystemGetter.get is marked async_helpers.exclusive, when we change the
source multiple times quickly, we create a "queue" of SystemGetter.get tasks.
This can take a long time to consume, and really we only need the result of the
last query.

When _examine_systems gets cancelled, we now check if the SystemGetter.get task
has successfully acquired the lock. If it hasn't, then we cancel it.

LP: #2084032

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>