storage: only if it makes sense, cancel _get_systems
Now that _get_systems is marked async_helpers.exclusive, when we change the
source multiple times quickly, we create a "queue" of _get_systems tasks.
This can take a long time to consume, and really we only need the result of the
last query.

When _examine_systems gets cancelled, we now check if the _get_systems task
has successfully acquired the lock. If it hasn't, then we cancel it.

LP: #2084032

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>
(cherry picked from commit 92184567c816bedb7def8d51681007740c8fea61)

Backported to use _get_system() instead of SystemGetter.get() from the original
patch.