async_helpers: add "started_event" to exclusive
When calling a coroutine function that is decorated with
"async_helpers.exclusive", one can now optionally pass an asyncio.Event.

If it is specified, the coroutine will set the event after it starts executing
the body of the coroutine (i.e. after it acquires the lock).

This is useful if the coroutine function is normally not safe to cancel in the
middle of its critical section. Using this mechanism, the parent can know if
the task can still be cancelled or if it is too late.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>