storage: always accept to resize a partition that is being wiped
If we are wiping a partition, don't bother checking if the filesystem that was
on it supports resizing. The filesystem is going to be erased anyway. The new
filesystem (if any) will be created using mkfs.* so there is no need for resize
support.

LP: #2096799

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>
(cherry picked from commit 850b8718dab8c7b7a64c84ac70631926959cb416)