storage: fix use of ptable parameter when calling reformat
When reformatting a disk using the following call:

  > manipulator.reformat(disk, ptable="something")

The ptable parameter would only appear to be honored if the disk did not
contain any partition. This happened because the reformat() function sets the
ptable of the disk first and then starts removing partitions using
delete_partition() in a loop.

However, delete_partition() has a mechanism to reset the partition table to
None when it is invoked on the last partition of a disk.

Therefore, the disk ended up with None as the partition table (ignoring the
ptable value) if the disk contained any partition.

Fixed by moving the assignment of disk.ptable after removing partitions.
Arguably, it is a mistake to change the partition table value before removing
the partitions anyway.

LP: #2097091

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>