storage: when calling reformat with ptable=None, use the default ptable type
Previously, the meaning of ptable=None as in reformat(disk, ptable=None) was to
create a new partition table of the existing type. Unfortunately, this was
only honored when disks didn't have partitions ; which is a rather uncommon
scenario.

For disks with existing partitions, the value of ptable was accidentally
ignored. The new partition table would always be created with the default type
(i.e., GPT on most platforms).

Before reformatting a disk with partitions is the most commons scenario, we are
changing the meaning of ptable=None when passed as an argument to reformat().
Instead of meaning "keep the current type", it now means "use the default
type". This should avoid changing the behavior for the most common scenario.

LP: #2097091

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>
(cherry picked from commit b1f2b3568f398fc81ca234fa71ff234cc17b3f94)