storage: make sure API consumers know if disk is modifiable without reformat
Disks that have an unsupported partition table will now be marked
"requires_reformat": true in the v2 storage API.

Clients should prevent modifications of a disk which is marked
"requires_reformat": true. This includes:

 * creating a new partition
 * removing a partition
 * modifying a partition
 * marking the disk as the boot device

Essentially, the only action that the client should support for such disk is to
reformat it.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>