ci: update cla-check to has-signed-canonical-cla@v2
The option for accepting existing contributors was removed
from the CLA checker in v2 with the following reasoning[1,2]:

	Remove the option for accepting existing contributors, as
	contributors can revoke their CLA agreement. The CLA must
	be checked with every commit.

For Subiquity's case, this seems to mostly have been used as a
workaround for Canonical employees signing commits with their Ubuntu
member emails addresses[3], which the v1 CLA checker had trouble with.
This doesn't look like it'll be an issue with v2 so we can safely proceed
with removing this.

[1] https://github.com/canonical/has-signed-canonical-cla/releases/tag/2.0.0
[2] https://github.com/canonical/has-signed-canonical-cla/commit/82ea789f7b1da38d1fbc82d92fd972d292bd453f
[3] 6b1700bd4766b6a77daf88842ea23172a40b22a2

(cherry picked from commit 4ce839b0dcc058d2fd2480177b4468e16cb52686)