storage: add (GPT) partition name in the API but don't allow to set it
The /storage/v2 endpoint will now return a "name" field for each partition. It
will either be a string or null if the partition does not have a name.

Currently, the name property of a partition is read-only. It cannot be changed
using the API.

The request handler for POST /storage/v2/edit_partition, one can omit the
"name" property entirely or repeat the value that the partition already had.

Examples:

 * Partition(number=1, name=None)
 * Partition(number=2, name="Microsoft Reserved Partition")

 POST /storage/v2/edit_partition
 > {..., "partition": {"number": 1, ...}}
     -> OK (not trying to change the name)

 POST /storage/v2/edit_partition
 > {..., "partition": {"number": 2, ...}}
     -> OK (not trying to change the name)

 POST /storage/v2/edit_partition
 > {..., "partition": {"number": 1, "name": null, ...}}
     -> OK (value of name hasn't changed)

 POST /storage/v2/edit_partition
 > {..., "partition": {"number": 1, "name": "something", ...}}
     -> KO (trying to change the value to "something")

LP: #2099893

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>