zdev: use command runner to benefit from dry-run implementation
Previously, in the Zdev controller, we had code to make sure that we didn't
call `chzdev` when running in dry-run mode. Instead we would call
asyncio.sleep.

We can make the code simpler though by using CommandRunner. It has a dry-run
implementation that does not actually execute commands - and allows configuring
a sleep duration.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>