shutdown: force to reboot/poweroff despite shutdown inhibits
Since systemd 257, logind also honors shutdown inhibits if the caller is root.
This prevents Subiquity from rebooting at the end of the installation on
plucky. On older releases of Ubuntu (with systemd < 257), shutdown inhibits
were ignored when the caller was root.

We now pass the --check-inhibitors=no option to ignore the inhibitor when
Subiquity requests a reboot.

LP: #2092438

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>