models: we don't need curtin's grub replace_linux_default=False
Curtin knows to carry forth kernel command line after ---.  We set
replace_linux_default=False in subiquity with the goal of having `quiet
splash` be in the default grub config, but `--- quiet splash` is already
specified in the ISO grub config.

This matters because on arm64 we need more parameters, so we want those
to flow through to the default grub config on the target system.

LP: #2104310