kvm-test: support carryover parameters
When supplying kernel cmdline to the VM, look at grub.cfg and, if the
triple dash is found, take the parameters from that and add that to our
kernel command line.

This is a more accurate form of testing, because under the current
implementation, the target system grub defaults have a different value
depending on if kvm-test was used with or without a kernel command line.

(implied by https://github.com/canonical/subiquity/pull/2188)