storage: fix display of potential devices for MD or LVM
When creating or editing an LVM or MD container, the list of potential devices
is shown. Unfortunately, when one of the devices has a label that is too long,
all of the devices used to become unvisible, making it challenging to create a
RAID / LVM properly.

Ensure, the list of devices is now visible no matter the length of labels.

LP: #2045688

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>