apt: remove /var/lib/apt/lists from apt runs
In testing against a country mirror, apt-get update may fail with:
  File has unexpected size (x != y). Mirror sync in progress?

On inspection, we found that the copy of /var/lib/apt/lists was in a bad
state and that further apt-get update invocation would not correct it.

By removing this cached information, we have to download more from the
mirror (not great), but the result should be more robust.

LP: #2105480