storage: make the backup location optional for recovery key
When using the LUKS flow in the TUI, a backup location for the recovery key is
required. However, in the desktop installer flow with TPM/FDE, the client
manages what to do with the recovery key. Therefore, we need to make the backup
location optional.

The backup location is still set when creating the RecoveryHandler object from
POST data, so the behavior when creating a recovery key in the TUI remains
unchanged.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>